[backup] restructure libradb::backup
Prepare for introducing the restore handler
